Revert "split unit / integration test"
This reverts commit 9e09c65ce5970f9d6b3857f5ffde01d9f2d36c56.
